+++ lib/td_auth_web/router.ex
@@ -36,11 +36,14 @@ defmodule TdAuthWeb.Router do
     get("/sessions", SessionController, :ping)
     delete("/sessions", SessionController, :destroy)
 
+    resources("/users/search", UserSearchController, only: [:create], singleton: true)
+
     resources "/users", UserController, except: [:new, :edit] do
       resources("/permissions", UserPermissionController, singleton: true, only: [:show], name: "permissions")
     end
 
     resources "/password", PasswordController, only: [:update], singleton: true
+    resources("/groups/search", GroupSearchController, only: [:create], singleton: true)
 
     resources("/groups", GroupController, except: [:new, :edit])