+++ test/td_auth_web/controllers/user_controller_test.exs
@@ -76,8 +76,8 @@ defmodule TdAuthWeb.UserControllerTest do
                |> json_response(:forbidden)
     end
 
-    test "user account can list all users if he has any permission in bg", %{conn: conn} do
-      {:ok, %{id: user_id, email: email, full_name: full_name, user_name: user_name} = user} =
+    test "user account cannot list all users even having any permission in bg", %{conn: conn} do
+      {:ok, user} =
         :user
         |> build(password: "pass000")
         |> Map.take([:user_name, :password, :email])
@@ -107,21 +107,10 @@ defmodule TdAuthWeb.UserControllerTest do
       assert {:ok, %{"groups" => ["business_glossary_view"]}} =
                Guardian.decode_and_verify(token, %{"typ" => "access"})
 
-      assert %{
-               "data" => [
-                 %{
-                   "email" => ^email,
-                   "full_name" => ^full_name,
-                   "id" => ^user_id,
-                   "role" => "user",
-                   "user_name" => ^user_name
-                 }
-               ]
-             } =
-               conn
+      assert conn
                |> put_auth_headers(token)
                |> get(Routes.user_path(conn, :index))
-               |> json_response(:ok)
+               |> json_response(:forbidden)
     end
   end